+++ Sources/TuistSupportTesting/Extensions/URLRequest+TestData.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public extension URLRequest {
+    static func test(url: URL = URL.test()) -> URLRequest {
+        URLRequest(url: url)
+    }
+}