+++ CHANGELOG.md
@@ -4,10 +4,16 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ## Next
 
+### Fixed
+
+- Don't try to delete a file if it doesn't exist https://github.com/tuist/tuist/pull/1177 by @pepibumur
+
 ### Added
 
 - Encrypt/decrypt command https://github.com/tuist/tuist/pull/1127 by @fortmarek
 - A link to the example app in the uFeatures documentation https://github.com/tuist/tuist/pull/1176 by @pepibumur.
+- Add ProjectGeneratorGraphMapping protocol and use it from ProjectGenerator https://github.com/tuist/tuist/pull/1178 by @pepibumur
+- `CloudSessionController` component to authenticate users https://github.com/tuist/tuist/pull/1174 by @pepibumur.
 
 ## 1.5.4